Added controller and more persistence.

+++ journeyrec-server/src/main/java/com/mottmac/journeyrec/dao/JourneyPointRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface JourneyPointRepository extends CrudRepository<JourneyPoint, Integer> {
+public interface JourneyPointRepository extends CrudRepository<JourneyPoint, Long> {
     List<JourneyPoint> fetchJourneyPoints();
 }
